Changes and notes for Initial Production
Changes for getting this to work in production in Apache on my host.
WSGI configuration proved a lot more complicated than expected.
