+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = dmdma
-version = 0.1.0
+version = 0.1.2
 url = https://github.com/d-roman-halliday/Dirty-Master-Data-Management-Application
 
 [options]
