+++ setup.py
@@ -2,7 +2,8 @@ from setuptools import setup
 
 setup(
     name='dmdma',
-    packages=['dmdma'],
+    # ToDo: Apply models from : https://flask.palletsprojects.com/en/2.1.x/patterns/distribute/?highlight=find_packages
+    packages=['dmdma','dmdma.reporting'],
     include_package_data=True,
     install_requires=[
         'flask',
